+++ src/steps/resource-manager/databases/sql/client.ts
@@ -2,6 +2,7 @@ import { SqlManagementClient } from '@azure/arm-sql';
 import {
   Database,
   DatabaseBlobAuditingPoliciesGetResponse,
+  EncryptionProtectorsGetResponse,
   FirewallRule,
   Server,
   ServerAzureADAdministrator,
@@ -90,6 +91,39 @@ export class SQLClient extends Client {
       resourceDescription: 'sql.server.activeDirectoryAdmins',
       callback,
     });
+  }
+
+  public async fetchServerEncryptionProtector(server: {
+    name: string;
+    id: string;
+  }): Promise<EncryptionProtectorsGetResponse | undefined> {
+    const serviceClient = await this.getAuthenticatedServiceClient(
+      SqlManagementClient,
+    );
+
+    const resourceGroup = resourceGroupName(server.id, true);
+    const serverName = server.name;
+
+    try {
+      const response = await serviceClient.encryptionProtectors.get(
+        resourceGroup,
+        serverName,
+      );
+      return response;
+    } catch (err) {
+      this.logger.warn(
+        {
+          err: new IntegrationProviderAPIError({
+            endpoint: 'sql.encryptionProtectors',
+            status: err.status,
+            statusText: err.statusText,
+            cause: err,
+          }),
+          server: server.id,
+        },
+        'Failed to obtain encryption protectors for server',
+      );
+    }
   }
 
   public async fetchDatabaseEncryption(