+++ src/steps/resource-manager/databases/sql/index.ts
@@ -1,5 +1,6 @@
 import {
   createDirectRelationship,
+  getRawData,
   IntegrationStepExecutionContext,
   RelationshipClass,
   Step,
@@ -17,6 +18,7 @@ import {
   setAuditingStatus,
   setDatabaseEncryption,
   setServerSecurityAlerting,
+  setServerEncryptionProtector,
 } from './converters';
 import createResourceGroupResourceRelationship from '../../utils/createResourceGroupResourceRelationship';
 import {
@@ -25,8 +27,9 @@ import {
   diagnosticSettingsRelationshipsForResource,
 } from '../../utils/createDiagnosticSettingsEntitiesAndRelationshipsForResource';
 import { STEP_RM_RESOURCES_RESOURCE_GROUPS } from '../../resources/constants';
+import { Server } from '@azure/arm-sql/esm/models';
 
-export async function fetchSQLDatabases(
+export async function fetchSQLServers(
   executionContext: IntegrationStepContext,
 ): Promise<void> {
   const { instance, logger, jobState } = executionContext;
@@ -49,6 +52,13 @@ export async function fetchSQLDatabases(
       serverEntity,
       await client.fetchServerSecurityAlertPolicies(server),
     );
+    setServerEncryptionProtector(
+      serverEntity,
+      await client.fetchServerEncryptionProtector({
+        name: server.name!,
+        id: server.id!,
+      }),
+    );
 
     await jobState.addEntity(serverEntity);
 
@@ -56,13 +66,39 @@ export async function fetchSQLDatabases(
       executionContext,
       serverEntity,
     );
+  });
+}
 
-    await createDiagnosticSettingsEntitiesAndRelationshipsForResource(
-      executionContext,
-      serverEntity,
-    );
+export async function fetchSQLServerDiagnosticSettings(
+  executionContext: IntegrationStepContext,
+): Promise<void> {
+  const { jobState } = executionContext;
 
-    try {
+  await jobState.iterateEntities(
+    { _type: entities.SERVER._type },
+    async (serverEntity) => {
+      await createDiagnosticSettingsEntitiesAndRelationshipsForResource(
+        executionContext,
+        serverEntity,
+      );
+    },
+  );
+}
+
+export async function fetchSQLDatabases(
+  executionContext: IntegrationStepContext,
+): Promise<void> {
+  const { instance, logger, jobState } = executionContext;
+  const accountEntity = await getAccountEntity(jobState);
+  const webLinker = createAzureWebLinker(accountEntity.defaultDomain as string);
+  const client = new SQLClient(instance.config, logger);
+
+  await jobState.iterateEntities(
+    { _type: entities.SERVER._type },
+    async (serverEntity) => {
+      const server = getRawData<Server>(serverEntity, 'default');
+      if (!server) return;
+
       await client.iterateDatabases(server, async (database) => {
         const databaseEntity = createDatabaseEntity(
           webLinker,
@@ -89,17 +125,8 @@ export async function fetchSQLDatabases(
           }),
         );
       });
-    } catch (err) {
-      logger.warn(
-        { err, server: { id: server.id, type: server.type } },
-        'Failure fetching databases for server',
-      );
-      // In the case this is a transient failure, ideally we would avoid
-      // deleting previously ingested databases for this server. That would
-      // require that we process each server independently, and have a way
-      // to communicate to the synchronizer that only a subset is partial.
-    }
-  });
+    },
+  );
 }
 
 export async function fetchSQLServerFirewallRules(
@@ -173,19 +200,27 @@ export const sqlSteps: Step<
   IntegrationStepExecutionContext<IntegrationConfig>
 >[] = [
   {
+    id: steps.SERVERS,
+    name: 'SQL Servers',
+    entities: [entities.SERVER],
+    relationships: [relationships.RESOURCE_GROUP_HAS_SQL_SERVER],
+    dependsOn: [STEP_AD_ACCOUNT, STEP_RM_RESOURCES_RESOURCE_GROUPS],
+    executionHandler: fetchSQLServers,
+  },
+  {
+    id: steps.SERVER_DIAGNOSTIC_SETTINGS,
+    name: 'SQL Server Diagnostic Settings',
+    entities: [...diagnosticSettingsEntitiesForResource],
+    relationships: [...diagnosticSettingsRelationshipsForResource],
+    dependsOn: [STEP_AD_ACCOUNT, steps.SERVERS],
+    executionHandler: fetchSQLServerDiagnosticSettings,
+  },
+  {
     id: steps.DATABASES,
     name: 'SQL Databases',
-    entities: [
-      entities.SERVER,
-      entities.DATABASE,
-      ...diagnosticSettingsEntitiesForResource,
-    ],
-    relationships: [
-      relationships.RESOURCE_GROUP_HAS_SQL_SERVER,
-      relationships.SQL_SERVER_HAS_SQL_DATABASE,
-      ...diagnosticSettingsRelationshipsForResource,
-    ],
-    dependsOn: [STEP_AD_ACCOUNT, STEP_RM_RESOURCES_RESOURCE_GROUPS],
+    entities: [entities.DATABASE],
+    relationships: [relationships.SQL_SERVER_HAS_SQL_DATABASE],
+    dependsOn: [STEP_AD_ACCOUNT, steps.SERVERS],
     executionHandler: fetchSQLDatabases,
   },
   {
@@ -193,7 +228,7 @@ export const sqlSteps: Step<
     name: 'SQL Server Firewall Rules',
     entities: [entities.FIREWALL_RULE],
     relationships: [relationships.SQL_SERVER_HAS_FIREWALL_RULE],
-    dependsOn: [steps.DATABASES],
+    dependsOn: [steps.SERVERS],
     executionHandler: fetchSQLServerFirewallRules,
   },
   {
@@ -201,7 +236,7 @@ export const sqlSteps: Step<
     name: 'SQL Server Active Directory Admins',
     entities: [entities.ACTIVE_DIRECTORY_ADMIN],
     relationships: [relationships.SQL_SERVER_HAS_AD_ADMIN],
-    dependsOn: [steps.DATABASES],
+    dependsOn: [steps.SERVERS],
     executionHandler: fetchSQLServerActiveDirectoryAdmins,
   },
 ];